added preprint
This paper introduces novel enhancements to the modified version of the BIRECT (BIsecting RECTangles) algorithm, denoted as BIRECTv. Our research explores diverse approaches, with a primary focus on incorporating a grouping strategy for hyper-rectangles of similar sizes. This categorization into different classes, constrained by a predefined threshold, aims to enhance computational efficiency, especially when dealing with a substantial number of hyper-rectangles of varying sizes. The introduction of this approach is particularly valuable in situations where efficient computation is essential. To further improve the algorithm's efficiency, we implemented a mechanism to prevent over-sampling and mitigate redundancy in sampling at shared vertices within descendant sub-regions. The use of a specific vertex database ensures a limit of two samples within each sub-region, contributing to more streamlined and effective optimization. Our experimental investigation demonstrates the positive impact of these improvements on the performance of the BIRECTv(imp.) algorithm. Comparisons with the original BIRECTv algorithm and its variants highlight the promising nature of our proposed algorithm as a global optimization solution. Furthermore, our results indicate that the BIRECTv(imp.) algorithm excels in addressing high-dimensional problems, showcasing its efficacy and versatility.

Signed-off-by: lchiter <77862332+lchiter@users.noreply.github.com>